feat: role

+++ src/routes/admin/role/index.tsx
@@ -1,49 +1,32 @@
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { Card } from 'antd';
+import { Card, Spin } from 'antd';
 import Column from 'antd/es/table/Column';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { useIntl } from 'react-intl';
+import { useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import { roleApi } from '../../../apis';
 import TableWrap from '../../../components/TableWrap';
 import CustomButton from '../../../components/buttons/CustomButton';
 import IconSVG from '../../../components/icons/icons';
 import CustomInput from '../../../components/input/CustomInput';
+import { ConfirmModel } from '../../../components/modals/ConfirmModel';
+import { ActionUser } from '../../../constants/enum';
 import { ADMIN_ROUTE_PATH } from '../../../constants/route';
-import { helper } from '../../../util/common';
-import { ActionUser, PERMISSIONS } from '../../../constants/enum';
-import { QUERY_LIST_ROLE } from '../../../util/contanst';
-import CheckPermission, { Permission } from '../../../util/check-permission';
-import { useSelector } from 'react-redux';
 import { RootState } from '../../../store';
+import { helper } from '../../../util/helper';
+import { QUERY_LIST_ROLE } from '../../../util/contanst';
 
 const ListRole = () => {
   const intl = useIntl();
   const navigate = useNavigate();
+  const queryClient = useQueryClient();
   const [page, setPage] = useState<number>(1);
   const [size, setSize] = useState<number>(10);
-  const [sort, setSort] = useState<string>('');
+  const [sort, setSort] = useState<string | undefined>(undefined);
   const [fullTextSearch, setFullTextSearch] = useState<string>('');
-  const [permission, setPermission] = useState<Permission>({
-    read: false,
-    create: false,
-    delete: false,
-    update: false,
-  });
-  const queryClient = useQueryClient();
-
+  const [isShowModal, setIsShowModal] = useState<{ id: string; name: string | undefined }>();
   const { authUser } = useSelector((state: RootState) => state.auth);
-
-  useEffect(() => {
-    if (authUser?.user?.roles) {
-      setPermission({
-        read: Boolean(CheckPermission(PERMISSIONS.ReadRole, authUser)),
-        create: Boolean(CheckPermission(PERMISSIONS.CreateRole, authUser)),
-        delete: Boolean(CheckPermission(PERMISSIONS.DeleteRole, authUser)),
-        update: Boolean(CheckPermission(PERMISSIONS.DeleteRole, authUser)),
-      });
-    }
-  }, [authUser]);
 
   const { data, isLoading } = useQuery({
     queryKey: [QUERY_LIST_ROLE, { page, size, sort, fullTextSearch }],
@@ -67,41 +50,28 @@ const ListRole = () => {
     navigate(helper.showDetail(id));
   };
 
-  const handleRoleDelete = (id: string) => {
-    DeleteRole(id);
+  const handleDeleteRole = (id: string) => {
+    if (isShowModal?.id) {
+      DeleteRole(isShowModal?.id);
+    }
   };
 
   return (
-    <Card id="role-management">
-      <div className="role-management__header">
-        <div className="role-management__header__title">
-          {intl.formatMessage({
-            id: 'role.list.title',
-          })}
+    <Spin spinning={isLoading}>
+      <Card>
+        <div className='d-flex justify-content-between align-items-center'>
+          <div className='font-weight-700 font-size-18 font-base'> {intl.formatMessage({ id: 'role.list.title' })}</div>
+          <CustomButton icon={<IconSVG type="create" />} onClick={() => { navigate(ADMIN_ROUTE_PATH.CREATE_ROLE) }}>
+            {intl.formatMessage({ id: 'common.create' })}
+          </CustomButton>
         </div>
-        <CustomButton
-          className="button-add"
-          disabled={!permission.create}
-          icon={<IconSVG type="create" />}
-          onClick={() => {
-            navigate(ADMIN_ROUTE_PATH.CREATE_ROLE);
-          }}
-        >
-          {intl.formatMessage({
-            id: 'role.list.button.add',
-          })}
-        </CustomButton>
-      </div>
-      <CustomInput
-        placeholder={intl.formatMessage({
-          id: 'role.list.search',
-        })}
-        prefix={<IconSVG type="search" />}
-        className="input-search"
-      />
-      {permission.read && (
+        <CustomInput
+          placeholder={intl.formatMessage({ id: 'common.search' })}
+          prefix={<IconSVG type="search" />}
+          className="w-44 mt-32"
+        />
         <TableWrap
-          className="custom-table"
+          className="custom-table mt-32"
           data={data?.data.content}
           isLoading={isLoading}
           page={page}
@@ -113,30 +83,29 @@ const ListRole = () => {
         >
           <Column
             title={intl.formatMessage({
-              id: 'role.list.table.code',
+              id: 'table.code',
             })}
             dataIndex="code"
             width={'15%'}
           />
           <Column
             title={intl.formatMessage({
-              id: 'role.list.table.role',
+              id: 'table.role',
             })}
             dataIndex="name"
           />
           <Column
             title={intl.formatMessage({
-              id: 'role.list.table.action',
+              id: 'table.action',
             })}
             dataIndex="action"
             width={'15%'}
             render={(_, record: any) => (
-              <div className="action-role">
-                <div onClick={() => handleRoleEdit(record.id)} className={permission.update ? '' : 'disable'}>
+              <div className="d-flex justify-content-center align-items-center gap-2">
+                <div onClick={() => handleRoleEdit(record.id)} className='pointer'>
                   <IconSVG type="edit" />
                 </div>
-                <span className="divider"></span>
-                <div onClick={() => () => handleRoleDelete(record.id)} className={permission.delete ? '' : 'disable'}>
+                <div onClick={() => setIsShowModal({ id: record.id, name: record.name })} className='pointer'>
                   <IconSVG type="delete" />
                 </div>
               </div>
@@ -144,8 +113,15 @@ const ListRole = () => {
             align="center"
           />
         </TableWrap>
-      )}
-    </Card>
+      </Card>
+      <ConfirmModel
+        visible={!!isShowModal?.id}
+        onSubmit={() => handleDeleteRole}
+        onClose={() => {
+          setIsShowModal(undefined);
+        }}
+      />
+    </Spin>
   );
 };
